Added exercises to tracker for realtime render
